+++ testing/test_model.py
@@ -12,7 +12,7 @@ from pydantic import ValidationError
 from xmlschema.validators.exceptions import XMLSchemaValidationError
 
 from ome_types import from_tiff, from_xml, model, to_xml
-from ome_types.schema import NS_OME, URI_OME, get_schema, to_xml_element
+from ome_types._xmlschema import NS_OME, URI_OME, get_schema, to_xml_element
 
 SHOULD_FAIL_READ = {
     # Some timestamps have negative years which datetime doesn't support.
